feat: add new card controller

+++ src/main/java/com/ipz/mba/repositories/CardTypeRepository.java
@@ -3,6 +3,6 @@ package com.ipz.mba.repositories;
 import com.ipz.mba.entities.CardTypeEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CardTypeRepository extends CrudRepository<CardTypeEntity, Integer> {
-    CardTypeEntity findCardTypeEntitiesById(int id);
+public interface CardTypeRepository extends CrudRepository<CardTypeEntity, Long> {
+    CardTypeEntity findCardTypeEntitiesById(Long id);
 }
